+++ app/src/main/java/com/example/koohestantest1/local_db/entity/Product.java
@@ -63,7 +63,7 @@ public class Product {
     public boolean Show;
 
     @ColumnInfo(name = "update_date")
-    public int UpdateDate;
+    public long UpdateDate;
 
     @ColumnInfo(name = "deleted")
     public boolean Deleted;
